+++ src/seller.ts
@@ -12,7 +12,7 @@ import { UnitCode } from './unitCode';
 // strict definition(2022-08-04~)
 export type ISellerMerchantReturnPolicy = Pick<
     IMerchantReturnPolicy,
-    'itemCondition' | 'typeOf' | 'merchantReturnDays' | 'restockingFee' | 'url'
+    'itemCondition' | 'typeOf' | 'merchantReturnDays' | 'restockingFee' | 'url' | 'applicablePaymentMethod'
 > & {
     restockingFee: IRestockingFee;
 };