+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - インボイスのtotalPaymentDue.valueを必須化
+- creativeWorkインターフェースを最適化
 
 ### Deprecated